float counts

+++ scn_fs/src/main/kotlin/SCDatasetExpression.kt
@@ -14,7 +14,7 @@ import java.nio.file.Path
 
 data class SCExpressionJson(
     val features: List<String>,
-    val featureCounts: Map<String, Int>?,
+    val featureCounts: Map<String, Float>?,
     val barcodes: List<String>,
     val totalCounts: List<Int>,
     val expType: ExpressionType,
